+++ aa_forum/__init__.py
@@ -4,5 +4,5 @@ A couple of variables to use throughout the app
 
 default_app_config: str = "aa_forum.apps.AaForumConfig"
 
-__version__ = "0.1.0-beta.8"
+__version__ = "0.1.0-beta.9"
 __title__ = "Forum"