+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - "Show all unread topics" and "Mark all topics as read" buttons above the category
   list in forum index
 - "Mark all topics as read" in unread topics view above and below board list
-
+- More logging
 
 
 ## [0.1.0-beta.8] - 2021-07-18