+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.1.0-beta.9] - Unreleased
+## [0.1.0-beta.9] - 2021-07-19
 
 ### Added